requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asgiref=3.8.1=pyhd8ed1ab_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
django=5.1.2=pyhd8ed1ab_0
google-api-core=2.22.0=pypi_0
google-api-python-client=2.151.0=pypi_0
google-auth=2.36.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.1=pypi_0
googleapis-common-protos=1.65.0=pypi_0
httplib2=0.22.0=pypi_0
idna=3.10=pypi_0
ld_impl_linux-64=2.43=h712a8e2_2
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libmpdec=4.0.0=h4bc722e_0
libsqlite=3.47.0=hadc24fc_1
libuuid=2.38.1=h0b41bf4_0
libzlib=1.3.1=hb9d3cd8_2
ncurses=6.5=he02047a_1
oauthlib=3.2.2=pypi_0
openssl=3.3.2=hb9d3cd8_0
pip=24.3.1=pyh145f28c_0
proto-plus=1.25.0=pypi_0
protobuf=5.28.3=pypi_0
psycopg2-binary=2.9.10=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pyparsing=3.2.0=pypi_0
python=3.13.0=h9ebbce0_100_cp313
python_abi=3.13=5_cp313
readline=8.2=h8228510_1
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
sqlparse=0.5.1=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
xz=5.2.6=h166bdaf_0
